+++ src/ztype/string_encode.rs
@@ -5,7 +5,7 @@ use crate::ztype::unsigned_bitsize;
 use super::varuint_encode::write_varsize;
 
 pub fn write_string(writer: &mut BitWriter, s: &str) {
-    write_varsize(writer, s.len() as u64);
+    write_varsize(writer, s.len() as u32);
 
     for c in s.chars() {
         writer.write_u8(c as u8, 8).expect("failed to write string");
